Address `bugprone-narrowing-conversions` warnings.
The warnings were:

```
/w/dolphin-memory-engine/Source/GUI/MemCopy/DlgCopy.cpp:99:53: warning: narrowing conversion from 'u32' (aka 'unsigned int') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   99 |                            .arg(Common::MEM1_START, Common::GetMEM1End() - 1);
      |                                                     ^
/w/dolphin-memory-engine/Source/GUI/MemCopy/DlgCopy.cpp:102:71: warning: narrowing conversion from 'u32' (aka 'unsigned int') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
  102 |           tr(" or between 0x%08X and 0x%08X").arg(Common::MEM2_START, Common::GetMEM2End() - 1));
      |                                                                       ^
/w/dolphin-memory-engine/Source/GUI/MemViewer/MemViewer.cpp:92:38: warning: narrowing conversion from 'u32' (aka 'unsigned int') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
   92 |     verticalScrollBar()->setRange(0, ((m_memViewEnd - m_memViewStart) / m_numColumns) - m_numRows);
      |                                      ^
/w/dolphin-memory-engine/Source/GUI/MemViewer/MemViewer.cpp:146:35: warning: narrowing conversion from 'unsigned int' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
  146 |     verticalScrollBar()->setValue(((address & 0xFFFFFFF0) - m_memViewStart) / m_numColumns);
      |                                   ^
/w/dolphin-memory-engine/Source/GUI/MemViewer/MemViewer.cpp:170:36: warning: narrowing conversion from 'u32' (aka 'unsigned int') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
  170 |   verticalScrollBar()->setRange(0, ((m_memViewEnd - m_memViewStart) / m_numColumns) - m_numRows);
      |                                    ^
/w/dolphin-memory-engine/Source/GUI/MemViewer/MemViewer.cpp:727:21: warning: narrowing conversion from 'int' to signed type 'char' is implementation-defined [bugprone-narrowing-conversions]
  727 |       byteToWrite = (selectedMemoryValue & static_cast<char>(0xF0)) | byteToWrite;
      |                     ^
/w/dolphin-memory-engine/Source/GUI/MemViewer/MemViewer.cpp:730:11: warning: narrowing conversion from 'int' to signed type 'char' is implementation-defined [bugprone-narrowing-conversions]
  730 |           (selectedMemoryValue & static_cast<char>(0x0F)) | static_cast<char>(byteToWrite << 4);
      |           ^
/w/dolphin-memory-engine/Source/GUI/MemViewer/MemViewer.cpp:852:16: warning: narrowing conversion from 'u32' (aka 'unsigned int') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
  852 |     int byte = (m_currentFirstAddress + i) & 0xF;
      |                ^
/w/dolphin-memory-engine/Source/GUI/Settings/DlgSettings.cpp:160:27: warning: narrowing conversion from 'u32' (aka 'unsigned int') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
  160 |   m_sldMEM1Size->setValue(SConfig::getInstance().getMEM1Size() / 1024 / 1024);
      |                           ^
/w/dolphin-memory-engine/Source/GUI/Settings/DlgSettings.cpp:161:27: warning: narrowing conversion from 'u32' (aka 'unsigned int') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions]
  161 |   m_sldMEM2Size->setValue(SConfig::getInstance().getMEM2Size() / 1024 / 1024);
      |                           ^
```

+++ Source/GUI/MemCopy/DlgCopy.cpp
@@ -96,10 +96,10 @@ bool DlgCopy::copyMemory()
   {
     QString errorMsg = tr("The address you entered is invalid, make sure it is an "
                           "hexadecimal number between 0x%08X and 0x%08X")
-                           .arg(Common::MEM1_START, Common::GetMEM1End() - 1);
+                           .arg(Common::MEM1_START, static_cast<int>(Common::GetMEM1End()) - 1);
     if (DolphinComm::DolphinAccessor::isMEM2Present())
-      errorMsg.append(
-          tr(" or between 0x%08X and 0x%08X").arg(Common::MEM2_START, Common::GetMEM2End() - 1));
+      errorMsg.append(tr(" or between 0x%08X and 0x%08X")
+                          .arg(Common::MEM2_START, static_cast<int>(Common::GetMEM2End()) - 1));
 
     errorBox = new QMessageBox(QMessageBox::Critical, tr("Invalid address"), errorMsg,
                                QMessageBox::Ok, nullptr);
